fix shit

+++ Assets/Scripts/Player.cs
@@ -14,7 +14,7 @@ public class Player : MonoBehaviour
     public float JumpHeight = 1;
     public float FeetWidth = 1;
     public float GroundTestLength = 0.1f;
-    public float lives = 3.0f;
+    public int lives = 3;
     Rigidbody2D rb;
 
     Transform feet;
@@ -31,29 +31,33 @@ public class Player : MonoBehaviour
     {
         //Jump
 
-        IsGrounded = Physics2D.BoxCast(feet.position, new Vector2(FeetWidth,0.001f),0,Vector2.down, GroundTestLength, LayerMask.GetMask("Ground"));
+        IsGrounded = Physics2D.BoxCast(feet.position, new Vector2(FeetWidth, 0.001f), 0, Vector2.down, GroundTestLength, LayerMask.GetMask("Ground"));
 
-        if (IsGrounded && Input.GetButtonDown("Jump")) {
+        if (IsGrounded && Input.GetButtonDown("Jump"))
+        {
             IsJumping = true;
             rb.velocity = new Vector2(rb.velocity.x, Mathf.Sqrt(JumpHeight * 2 * Gravity));
         }
-        if (Input.GetButtonUp("Jump") && IsJumping) {
+        if (Input.GetButtonUp("Jump") && IsJumping)
+        {
             IsJumping = false;
             rb.velocity = new Vector2(rb.velocity.x, 0);
         }
 
-        if (rb.velocity.y > 0) {
+        if (rb.velocity.y > 0)
+        {
             rb.velocity += Vector2.down * Gravity * Time.deltaTime;
         }
-        else {
+        else
+        {
             IsJumping = false;
             rb.velocity += Vector2.down * Gravity * FallMultiplier * Time.deltaTime;
         }
 
         //Horizontal Movement
-        
+
         rb.velocity = new Vector2(GoingRight ? Speed : -Speed, rb.velocity.y);
-        
+
     }
 
 }
